refactor: Pass NCCH to RomFS instead of update method
A generic 'update' method doesn't seem flexible enough because we may need to change different parts of an object depending on which other parts of other objects are updated. Since it seems likely we'll need to do this, it makes more sense to pass the whole object.

Also, it makes sense to have each property within a class update its dependent properties.