fix: Fix NCCH title/program ID getter and setter
TypedArray.reverse() (like Array.reverse()) reverses the array in-place, in this case changing the actual data in the file, which we don't want. So store the Uint8Array in a class property and return a copy of it for the getter. For the setter, copy the input value, reverse it, and then store that.

Also add some more properties we're doing to need to write, including several more locations where it seems the title ID is being stored