feat: Update signatures when properties are changed
It might be less efficient to recalculate the signatures multiple times, but in the end it seems to be a pretty quick process. The alternative would be to have to keep track in each property of the dependencies that need to be updated, which is way too likely to result in something being missed and would be a headache to maintain.

I think from a big picture perspective, many other parts of the process will end up taking longer (downloading core files, or shifting the contents of the CIA file to insert a ROM) and so I think the added time inefficiency is probably not even going to matter in the end. Premature optimization and all that ;)